View function added

+++ src/main/resources/site/parts/student-list/student-list.js
@@ -4,17 +4,16 @@ var httpclient = require('/lib/http-client');
 var constants = require('../../../assets/constants/main.js');
 
 exports.get = function (req) {
-    log.info(req.params.response);
+
     var response = httpclient.request({
         method: "GET",
         url: `${constants.server}/students`,
         contentType: "application/json",
     });
     var students = JSON.parse(response.body);
-
     var model = {
         students: students,
-        message:req.params.response,
+        message: req.params.response,
     };
     var view = resolve('student-list.html');
 
@@ -22,28 +21,52 @@ exports.get = function (req) {
         body: thymeleaf.render(view, model)
     };
 
-
-
 };
 exports.post = function (req) {
 
-    var studentId=req.params.id;
+    if (req.params.button == 'Delete') {
 
-    var response = httpclient.request({
-        method: "DELETE",
-        url: `${constants.server}/students/${studentId}`,
-        contentType: "application/json",
+        var studentId = req.params.id;
+        var response = httpclient.request({
+            method: "DELETE",
+            url: `${constants.server}/students/${studentId}`,
+            contentType: "application/json",
 
-    });
-    var res=JSON.parse(response.body);
-    var url = portal.pageUrl({
-        path: `../${constants.home}`,
-        params:{
-            response:res.message
-        }
-    });
-    return {
+        });
+        var res = JSON.parse(response.body);
+        var url = portal.pageUrl({
+            path: `../${constants.home}`,
+            params: {
+                response: res.message
+            }
+        });
 
-    redirect: url,
-    };
+        return {
+
+            redirect: url,
+        };
+
+    }
+
+    if (req.params.button == 'View') {
+
+        var url = portal.pageUrl({
+            path: `../${constants.register}`,
+            params: {
+                id: req.params.id,
+                firstName: req.params.firstName,
+                lastName: req.params.lastName,
+                email: req.params.email,
+                gender: req.params.gender,
+                dob: req.params.dob,
+                teacher: req.params.teacher,
+            },
+        });
+
+        return {
+            redirect: url,
+        };
+
+    }
+
 };
